requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
wheel==0.24.0

